[Dev] Mega, the redeem tx is now working on testnet

+++ src/constants.rs
@@ -16,4 +16,4 @@ pub const GRIN_BLOCK_TIME :u32 = 1;
 pub const TEST_NET : bool = true;
 // The default value for sequence in bitcoin transactions
 pub const FFFFFFFF : u32 = 4294967295;
-pub const SIGHASH_ALL : u32 = 0x01;
+pub const SIGHASH_ALL : u8 = 0x01;
